azure_service/requirements.txt
azure-storage>=0.34.0,<0.35.0
pandas>=0.20.0,<0.21.0
